pyproject.toml
[tool.poetry]
name = "sentiment_lm"
version = "0.1.0"
description = ""
authors = ["Gabriel keith <gabe.keith@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jax = {extras = ["cpu"], version = "^0.4.28"}
flax = "^0.8.4"
optax = "^0.2.2"
orbax-checkpoint = "^0.5.15"
numpy = "^1.26.4"
matplotlib = "^3.9.0"
tensorboardx = "^2.6.2.2"
crc32c = "^2.4"
#tensorflow = "^2.16.1"
#jupyter = "^1.0.0"
#penzai = {extras = ["notebook"], version = "^0.1.2"}
pydantic = "^2.7.4"
tokenmonster = "^1.1.12"
gradio = "^4.41.0"

[tool.poetry.scripts]
sentiment_lm = "sentiment_lm.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
